moved raw changes classes
