+++ pmp-core/src/main/java/dk/nykredit/pmp/core/commit/ChangeFactory.java
@@ -1,5 +1,10 @@
 package dk.nykredit.pmp.core.commit;
 
+import dk.nykredit.pmp.core.remote.json.raw_types.RawChange;
+import dk.nykredit.pmp.core.remote.json.raw_types.RawCommitRevert;
+import dk.nykredit.pmp.core.remote.json.raw_types.RawParameterChange;
+import dk.nykredit.pmp.core.remote.json.raw_types.RawParameterRevert;
+
 public interface ChangeFactory {
     Change createChange(RawChange rawChange);
     ParameterChange createChange(RawParameterChange rawChange);
