+++ pmp-core/src/main/java/dk/nykredit/pmp/core/commit/ChangeFactoryImpl.java
@@ -1,5 +1,10 @@
 package dk.nykredit.pmp.core.commit;
 
+import dk.nykredit.pmp.core.remote.json.raw_types.RawChange;
+import dk.nykredit.pmp.core.remote.json.raw_types.RawCommitRevert;
+import dk.nykredit.pmp.core.remote.json.raw_types.RawParameterChange;
+import dk.nykredit.pmp.core.remote.json.raw_types.RawParameterRevert;
+
 public class ChangeFactoryImpl implements ChangeFactory{
 
     public Change createChange(RawChange rawChange) {
