+++ pmp-core/src/main/java/dk/nykredit/pmp/core/commit/CommitFactory.java
@@ -1,5 +1,7 @@
 package dk.nykredit.pmp.core.commit;
 
+import dk.nykredit.pmp.core.remote.json.raw_types.RawCommit;
+
 public interface CommitFactory {
     Commit createCommit(RawCommit rawCommit);
 }
