+++ pmp-core/src/main/java/dk/nykredit/pmp/core/remote/json/raw_types/RawChange.java
@@ -1,4 +1,4 @@
-package dk.nykredit.pmp.core.commit;
+package dk.nykredit.pmp.core.remote.json.raw_types;
 // This is a marker interface, it is meant to be empty.
 public interface RawChange {
 }
