+++ pmp-core/src/main/java/dk/nykredit/pmp/core/remote/json/raw_types/RawParameterRevert.java
@@ -1,5 +1,6 @@
-package dk.nykredit.pmp.core.commit;
+package dk.nykredit.pmp.core.remote.json.raw_types;
 
+import dk.nykredit.pmp.core.commit.Service;
 import lombok.Getter;
 import lombok.Setter;
