+++ pmp-core/src/main/java/dk/nykredit/pmp/core/remote/servlet/CommitServlet.java
@@ -12,11 +12,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import dk.nykredit.pmp.core.commit.Commit;
 import dk.nykredit.pmp.core.commit.CommitDirector;
 import dk.nykredit.pmp.core.commit.CommitFactory;
-import dk.nykredit.pmp.core.commit.RawCommit;
 import dk.nykredit.pmp.core.commit.exception.OldValueInconsistentException;
 import dk.nykredit.pmp.core.commit.exception.StoredValueNullException;
 import dk.nykredit.pmp.core.commit.exception.TypeInconsistentException;
 import dk.nykredit.pmp.core.remote.json.ObjectMapperFactory;
+import dk.nykredit.pmp.core.remote.json.raw_types.RawCommit;
 
 public class CommitServlet extends HttpServlet {
